Add -Wall to find out about more warnings
Some of these fixes are temporary hacks that need deeper thought. For
example, what should we *really* do if we can't write back to the
accepted file descriptor? Close the connection? Attempt retries?

Warnings categories fixed:
* unused variable ‘XYZ’ [-Wunused-variable]
* ignoring return value of ‘XYZ’, declared with attribute warn_unused_result [-Wunused-result]
* control reaches end of non-void function [-Wreturn-type]

Some of these warnings must have been the default on Linux, because I
saw them come through on the Linux VM but not on the OSX side. I wanted
to know what compiler Linux was using when I say $(CC) in make, and to
do that I used strace:

    strace -f -o/tmp/compile.log make

The `-f` says to trace subprocesses (which is super-important because
I don't care what the fork/exec'ing situation is in the build, just what
ends up happening at some point when the compiler gets exec'ed). And the
`-o` lets me specify a path to the file.

I remembered reading in @jvns's strace zine about the `-e` option (which
lets you specify what system calls to trace), but I started with
`-etrace=open`, thinking the file would have to be opened for reading
before being exec'ed. That was a dead end, because the compiler is an
ELF executable (not some silly dynamic-language script like I use all
the time everywhere). So it never gets opened, only exec'ed. Next
mistake: `-etrace=exec`. This was just wrong off the bat: "invalid
system call 'exec'" was the error message. As I found out when I opened
the firehose by leaving off the `-e` option altogether, the actual
system call name is `execve`, so `-etrace=execve` was what I was looking
for originally.

Another thing I learned here is that you can't pipe strace output
directly to `less` (or anything else), because it prints to stderr. But
that's easy enough to fix by redirecting stderr to stdout:

    strace -f -etrace=execve make 2>&1 | less