Extract a test function for later use
Also improves reporting by including the test name in both failures and
passing tests. Not everything should be this high-level of a test, and
eventually I'll want to get ahold of a C unit testing setup, but for now
it feels OK to hack around learning the language and APIs with a sanity
check or two making sure the happy path still works.

This extracted test function will be helpful for IPv6 addresses. I'm
currently not able to support those due to inet_ntoa and inet_aton,
which it turns out are both considered obsolete since they only work
with IPv4.

So coming up soon will be cutting over to getaddrinfo / getnameinfo for
going back & forth between strings & sockaddrs.