Use the C11 standard
Initially got a new page full of errors & warnings when enabling this
(or c99) - this turned out to be because <netdb.h> (under Linux) only
defines the addrinfo struct when __USE_POSIX is defined, and it only
defines NI_MAXHOST/NI_MAXSERV when __USE_MISC is defined. `#define`'ing
those two directly before the `#include <netdb.h>` works, but it seemed
pretty gross (especially since things already looked good on OSX).

A little googling suggests that __USE_MISC, in particular, is more of an
internal flag, and it's better to set it indirectly via user-facing
constant definitions like _GNU_SOURCE. Defining that in code also gets
things working, but felt dirty - and more googling confirmed that it's
common to set this kind of thing as a flag to the compiler instead of
putting it directly into the code.

Learned about a very basic compiler flag:
* -D: allows `#define`'ing outside of source code. In this case,
  _GNU_SOURCE works on both Linux and OSX, but future features could
  potentially need e.g. _BSD_SOURCE on the OSX side, which will work for
  this commit's purposes as well (but doesn't work on Linux). The point
  is that if we needed to be able to support both platforms and for
  whatever reason _GNU_SOURCE didn't work on OSX, we could have a
  separate build script instead of monkeying around with the source
  code.