Move echo server build to its own makefile
After a couple of weeks studying James Grenning's TDD for Embedded C and
working along with his code examples and build process, it's time to get
that good stuff in here. So for now it's using CPPUnit's build process
in a pretty janky way. No unit tests yet, still leaning on the
acceptance level, and providing my own main inside the production-code
library, which won't fly at all when it comes time for CPPUnit to use
its main.

There were a few new compiler warnings to be fixed: functions with
neither previous declarations in header files (needing to be made
static), and a needing to add a `noreturn` attribute for an infinite
loop.